+++ packages/taro-transformer-wx/src/jsx.ts
@@ -38,8 +38,10 @@ export function removeJSXThisProperty (path: NodePath<t.ThisExpression>) {
   }
 }
 
-export function findJSXAttrByName (attrs: t.JSXAttribute[], name: string) {
+export function findJSXAttrByName (attrs: (t.JSXAttribute | t.JSXSpreadAttribute)[], name: string) {
   for (const attr of attrs) {
+    if (!t.isJSXAttribute(attr)) continue
+
     if (!t.isJSXIdentifier(attr.name)) {
       break
     }
@@ -279,7 +281,9 @@ export function parseJSXElement (element: t.JSXElement, isFirstEmit = false): st
         if (componentTransfromProps && componentTransfromProps[componentName]) {
           const transfromProps = componentTransfromProps[componentName]
           Object.keys(transfromProps).forEach(oriName => {
-            name = transfromProps[oriName]
+            if (name === oriName) {
+              name = transfromProps[oriName]
+            }
           })
         }
         if ((componentName === 'Input' || componentName === 'input') && name === 'maxLength') {