+++ packages/taro-transformer-wx/src/options.ts
@@ -38,6 +38,7 @@ export const buildBabelTransformOptions: () => TransformOptions = () => {
     plugins.push(buildVistor())
   }
   return {
+    filename: transformOptions.sourcePath,
     parserOpts: {
       sourceType: 'module',
       plugins: [
@@ -59,6 +60,6 @@ export const buildBabelTransformOptions: () => TransformOptions = () => {
     plugins: plugins
       .concat(functionalComponent)
       .concat(process.env.ESLINT === 'false' || transformOptions.isNormal || transformOptions.isTyped ? [] : eslintValidation)
-      .concat((isTestEnv) ? [] : require('babel-plugin-remove-dead-code').default)
+      .concat((isTestEnv) ? [] : require('babel-plugin-minify-dead-code').default)
   }
 }