+++ packages/taro/types/api/device/device.battery.d.ts
@@ -1,0 +1,4 @@
+declare namespace Taro {
+  // TODO: wx.getBatteryInfoSync
+  // TODO: wx.getBatteryInfo
+}